% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_Brownian_motion.R
\name{geometric_Brownian_motion}
\alias{geometric_Brownian_motion}
\title{Geometric Brownian motion}
\usage{
geometric_Brownian_motion(T, dt, mu, sigma, Y0)
}
\arguments{
\item{T}{total time}

\item{dt}{time step}

\item{mu}{mean}

\item{sigma}{standard deviation of the stochastic term, over 1 time unit}

\item{Y0}{initial value}
}
\value{
The time and the process realization vectors
\itemize{
  \item 1 - The time vector t.
  \item 2 - Tbe process realization Y.
}
}
\description{
Simulates a stochastic realization of geometric Brownian motion
The geometric Brownian motion is often taken as the standard model of stocks and other financial instruments.
It is also one of the building blocks of the Black-Scholes option pricing model.
}
