% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_Ornstein_Uhlenbeck.R
\name{geometric_Ornstein_Uhlenbeck}
\alias{geometric_Ornstein_Uhlenbeck}
\title{Geometric Ornstein-Uhlenbeck process}
\usage{
geometric_Ornstein_Uhlenbeck(T, dt, mu, sigma, tau, Y0)
}
\arguments{
\item{T}{total time}

\item{dt}{time step}

\item{mu}{mean}

\item{sigma}{standard deviation of the stochastic term, over 1 time unit}

\item{tau}{reverting time constant}

\item{Y0}{initial value}
}
\value{
The time and the process realization vectors
\itemize{
  \item 1 - The time vector t.
  \item 2 - Tbe process realization Y.
}
}
\description{
Simulates a geometric Ornstein-Uhlenbeck stochastic realization
The geometric Ornstein-Uhlenbeck process, like its classic version, models mean-reverting processes.
}
