% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ornstein_Uhlenbeck.R
\name{Ornstein_Uhlenbeck}
\alias{Ornstein_Uhlenbeck}
\title{Ornstein-Uhlenbeck process}
\usage{
Ornstein_Uhlenbeck(T, dt, mu, sigma, tau, Y0)
}
\arguments{
\item{T}{total time}

\item{dt}{time step}

\item{mu}{mean}

\item{sigma}{standard deviation of the stochastic term, over 1 time unit}

\item{tau}{reverting time constant}

\item{Y0}{initial value}
}
\value{
The time and the process realization vectors
\itemize{
  \item 1 - The time vector t.
  \item 2 - Tbe process realization Y.
}
}
\description{
Simulates an Ornstein-Uhlenbeck stochastic realization
The Ornstein-Uhlenbeck process is the canonical model of a mean-reverting process.
It is often used to model the spread in a pairs trading strategy.
}
